Minor refactoring to fix failing tests
